Change thalamus region list regex
This updates the regular expression used for thalamus placement-hints to
be in a different format that has been tested successfully, excludes
habenular and peripeduncular subregions, and to be valid for the
hierarchy/annotation used at its appropriate step in the Atlas pipeline.
For information on which regions were chosen and this list was created,
see the internal BBP Confluence page located at "Circuits > Mouse
Thalamus > Atlas-based Whole-thalamus subregion selection". This regex
has been built from the region list of the desired and present thalamus
regions as of the "final" version of the hierarchy and annotation built
by the Atlas pipeline, which is the output of the rule
`split_barrel_ccfv3_l23split`.

This change is meant to go in tandem with
https://github.com/BlueBrain/atlas-direction-vectors/pull/27 .