Fix final linting issues
This does a lot of small things for passing the linting.

For mypy, I had to add additional ignores for the Trimesh returned
types, since the ignore on the module as a whole wasn't preventing mypy
from expecting `load_mesh` to return a Geometry object, which is a
grandparent of Trimesh objects. I don't know if Trimesh changed their
API, I couldn't figure it out from the docs, and I don't know why mypy
was raising this now. In all the cases I could test or see, a proper
"Trimesh" object was returned instead of the more generic Geometry. I
don't think we need to worry about this.

For the pylint disable W0613 (unused-argument), I needed some
polymorphism for the thalamus case, but I wasn't sure how to handle that
alongside the linters' type-checking. I think this is the simplest
solution and is harmless.

Everything else is minor.