Update thalamus (#11)
* Overhaul the thalamus into steps, and add docs

Load thalamus meshes from CLI input, clean CLI

* Improve Blender instructions

* Change thalamus region list regex

This updates the regular expression used for thalamus placement-hints to
be in a different format that has been tested successfully, excludes
habenular and peripeduncular subregions, and to be valid for the
hierarchy/annotation used at its appropriate step in the Atlas pipeline.
For information on which regions were chosen and this list was created,
see the internal BBP Confluence page located at "Circuits > Mouse
Thalamus > Atlas-based Whole-thalamus subregion selection". This regex
has been built from the region list of the desired and present thalamus
regions as of the "final" version of the hierarchy and annotation built
by the Atlas pipeline, which is the output of the rule
`split_barrel_ccfv3_l23split`.

This change is meant to go in tandem with
https://github.com/BlueBrain/atlas-direction-vectors/pull/27 .

* Update layer names to be Atlas-Pipeline-compatible

* Fix formatting errors

* Attempt to update tests for new thal workflow

* Replace part of test anno with region in metadata

* fix test + format

* format

* Fix final linting issues

This does a lot of small things for passing the linting.

For mypy, I had to add additional ignores for the Trimesh returned
types, since the ignore on the module as a whole wasn't preventing mypy
from expecting `load_mesh` to return a Geometry object, which is a
grandparent of Trimesh objects. I don't know if Trimesh changed their
API, I couldn't figure it out from the docs, and I don't know why mypy
was raising this now. In all the cases I could test or see, a proper
"Trimesh" object was returned instead of the more generic Geometry. I
don't think we need to worry about this.

For the pylint disable W0613 (unused-argument), I needed some
polymorphism for the thalamus case, but I wasn't sure how to handle that
alongside the linters' type-checking. I think this is the simplest
solution and is harmless.

Everything else is minor.

* Make Alexis changes to CLI

* Apply MG code review changes

---------

Co-authored-by: Austin E. Soplata <austin.soplata@epfl.ch>
Co-authored-by: arnaudon <alexis.arnaudon@epfl.ch>